# openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "FastAPI",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/productor/": {
            "get": {
                "tags": [
                    "productor"
                ],
                "summary": "List Productors",
                "operationId": "list_productors_api_v1_productor__get",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 0,
                            "title": "Offset"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "default": 100,
                            "title": "Limit"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Productor"
                                    },
                                    "title": "Response List Productors Api V1 Productor  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "productor"
                ],
                "summary": "Create Productor",
                "operationId": "create_productor_api_v1_productor__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductorCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Productor"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/productor/{productor_id}": {
            "get": {
                "tags": [
                    "productor"
                ],
                "summary": "Get Productor",
                "operationId": "get_productor_api_v1_productor__productor_id__get",
                "parameters": [
                    {
                        "name": "productor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Productor Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Productor"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "productor"
                ],
                "summary": "Update Productor",
                "operationId": "update_productor_api_v1_productor__productor_id__put",
                "parameters": [
                    {
                        "name": "productor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Productor Id"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProductorCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Productor"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "productor"
                ],
                "summary": "Delete Productor",
                "operationId": "delete_productor_api_v1_productor__productor_id__delete",
                "parameters": [
                    {
                        "name": "productor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Productor Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Productor"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/productor/{productor_id}/farms": {
            "get": {
                "tags": [
                    "productor"
                ],
                "summary": "Get Productor Farms",
                "operationId": "get_productor_farms_api_v1_productor__productor_id__farms_get",
                "parameters": [
                    {
                        "name": "productor_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "title": "Productor Id"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Farm"
                                    },
                                    "title": "Response Get Productor Farms Api V1 Productor  Productor Id  Farms Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Farm": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "city": {
                        "type": "string",
                        "title": "City"
                    },
                    "state": {
                        "type": "string",
                        "title": "State"
                    },
                    "total_area": {
                        "type": "number",
                        "title": "Total Area"
                    },
                    "arable_area": {
                        "type": "number",
                        "title": "Arable Area"
                    },
                    "vegetation_area": {
                        "type": "number",
                        "title": "Vegetation Area"
                    },
                    "productor_id": {
                        "type": "integer",
                        "title": "Productor Id"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "city",
                    "state",
                    "total_area",
                    "arable_area",
                    "vegetation_area",
                    "productor_id",
                    "id"
                ],
                "title": "Farm"
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "Productor": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "cpf_cnpj": {
                        "type": "string",
                        "title": "Cpf Cnpj"
                    },
                    "birthdate": {
                        "type": "string",
                        "format": "date",
                        "title": "Birthdate"
                    },
                    "id": {
                        "type": "integer",
                        "title": "Id"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "cpf_cnpj",
                    "birthdate",
                    "id"
                ],
                "title": "Productor"
            },
            "ProductorCreate": {
                "properties": {
                    "name": {
                        "type": "string",
                        "title": "Name"
                    },
                    "cpf_cnpj": {
                        "type": "string",
                        "title": "Cpf Cnpj"
                    },
                    "birthdate": {
                        "type": "string",
                        "format": "date",
                        "title": "Birthdate"
                    }
                },
                "type": "object",
                "required": [
                    "name",
                    "cpf_cnpj",
                    "birthdate"
                ],
                "title": "ProductorCreate"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        }
    }
}